% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.test_expl.R
\name{ks.test_expl}
\alias{ks.test_expl}
\title{simulation and plots explaining KS-test and Wilcox-test}
\usage{
ks.test_expl(N_other = 1000, N_set = 100, alternative = "greater",
  z_score = 1.5, prop_pos = 0.3, prop_neg = 0, x_coor_other = -1.7,
  x_coor_set = 1.7, cex = 1.7,
  title = "Treatment: anti-Oct4 shRNA.\\nOct4 is a pos. and a neg. regulator of its targets",
  xlab = "z-score, simulated", lab_other = "other genes",
  lab_set = "Oct4 targets", seed = 1, x = NULL, y = NULL,
  gener_dist = NULL, gener_dist_lab = "all genes", distr = rnorm,
  col_other = "black", col_set = "blue", ...)
}
\arguments{
\item{N_other}{integer, size of the "other" group, ignored if x is provided}

\item{N_set}{integer, size of the "set" group, ignored if y is provided}

\item{alternative}{character, "greater" or "less". The choice of alternative hypothesis for KS-test \code{\link[stats]{ks.test}}. Greater tests is the set (y) distribution (or, in case of Wilcox test, it's median) is shifted to the right compated to other (x) distribution. This choice of "greater" or "less" is flipped for the Wilcox test due to difference in the null hypothesis formulation (\code{\link[stats]{wilcox.test}}).}

\item{z_score}{numeric, size of the true difference}

\item{prop_pos}{numeric, between 0 and 1, the fraction of "set" group that is higher than other by \code{z_score}}

\item{prop_neg}{numeric, between 0 and 1, the fraction of "set" group that is lower than other by \code{z_score}}

\item{x_coor_other}{numeric, x coordinate of "other" group label on the plot}

\item{x_coor_set}{numeric, x coordinate of "set" group label on the plot}

\item{cex}{numeric, plotting parameter, size of the group label text, axis label text and axis annotation text}

\item{title}{character, title of the plot}

\item{xlab}{character, x-axis label}

\item{lab_other}{character, other group label}

\item{lab_set}{character, set group label}

\item{seed}{integer, to initialise random number generator for reproducibility}

\item{x}{numeric vector, the distribution of values of the "other" group. These can come from the experimental data.}

\item{y}{numeric vector, the distribution of values of the "set" group. These can come from the experimental data.}

\item{distr}{function that generates a vector of random numbers of lenght n (n argument to this function). rnorm by default}

\item{col_other}{color of the other distribution (passed to plot(col = )). Details \code{\link[graphics]{plot}}}

\item{col_set}{color of the other distribution (passed to plot(col = )). Details \code{\link[graphics]{plot}}}

\item{...}{other plotting parameters \code{\link[graphics]{plot}}}
}
\value{
list containing the output of KS-test and Wilcox-test
}
\description{
This function generates random data (rnorm, or other provided by distr argument) or uses user-provided data to explain how KS-test and Wilcox-test measure difference in distributions. It shows empirical cumulative distribution of z-scores for a given group of genes (set) and for all other genes (other). User can compare this plot to the KS-test and the Wilcox-test results.
}
\examples{
# Let's have a look at a TF that regulates
# some of it's targets positively, some negatively and some are not affected.
ks.test_expl(alternative = "greater", prop_pos = 0.33, prop_neg = 0.4, seed = 1, cex.main = 1.4)
}
\author{
Vitalii Kleshchevnikov
}
