% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCMap.R
\name{openCellInfo}
\alias{openCellInfo}
\title{Read Connectivity Map project cell line data (2nd phase, Level 5)}
\usage{
openCellInfo(CMap_files)
}
\arguments{
\item{CMap_files}{a list of directories and urls produced by \code{\link{loadCMap}}}
}
\value{
data.table containing Cell line Information from the Connectivity map project
}
\description{
Read Connectivity Map project cell line data (2nd phase, Level 5)
}
\author{
Vitalii Kleshchevnikov
}
