% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCMap.R
\name{openFeatureData}
\alias{openFeatureData}
\title{Read Connectivity Map project featureData (2nd phase, Level 5)}
\usage{
openFeatureData(CMap_files)
}
\arguments{
\item{CMap_files}{a list of directories and urls produced by \code{\link{loadCMap}}}
}
\value{
data.table containing featureData (\code{\link[Biobase]{featureData}}) from the Connectivity map project
}
\description{
Read Connectivity Map project featureData (2nd phase, Level 5)
}
\author{
Vitalii Kleshchevnikov
}
