% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCMap.R
\name{openpData}
\alias{openpData}
\title{Read Connectivity Map project sample descriptions (pData) (2nd phase, Level 5)}
\usage{
openpData(CMap_files)
}
\arguments{
\item{CMap_files}{a list of directories and urls produced by \code{\link{loadCMap}}}
}
\value{
data.table containing the sample descriptions (pData) (\code{\link[Biobase]{pData}}) from the Connectivity map project
}
\description{
Read Connectivity Map project sample descriptions (pData) (2nd phase, Level 5)
}
\author{
Vitalii Kleshchevnikov
}
