% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbTable.R
\name{perturbTable}
\alias{perturbTable}
\title{Summarise available perturbation data annotations using base::table}
\usage{
perturbTable(CMap_dirs, formula = pert_type ~ pert_time, pert_ids = NULL,
  pdata = NULL)
}
\arguments{
\item{formula}{specify which 2 columns to call \code{\link[base]{table}} on: \code{pert_type ~ pert_time} or \code{type ~ pert_type} to tabulate single column}

\item{pert_ids}{data.table (optional) a way to select a subset of perturbation data annotations. data.table should be the output of \code{\link{geneName2PerturbAnno}} or \code{\link{geneID2PerturbAnno}} (should contain distil_id column, a unique perturbation ID)}

\item{pdata}{(optional) data.table, the output of \code{\link{openpData}}. By default \code{perturbTable} reads Cmap file (path provided in CMap_dirs), for multiple test it can be useful to supply data.table.}

\item{CMap_files}{a list of directories and urls produced by \code{\link{loadCMap}}}
}
\value{
data.table containing the perturbation details from the Connectivity map project
}
\description{
\code{perturbTable} Cross Tabulates Connectivity map perturbation data annotations on columns of interest and subsets of perturbations
}
\seealso{
\code{\link{openCellInfo}}, \code{\link{loadCMap}}, \code{\link{geneName2PerturbAnno}}
}
\author{
Vitalii Kleshchevnikov
}
