% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCMAP.R
\name{readCMAP}
\alias{readCMAP}
\title{Read Connectivity Map data given perturbation annotations}
\usage{
readCMAP(PerturbAnno, CMap_files, landmark_only = F)
}
\arguments{
\item{PerturbAnno}{data.table containing the perturbation details from the Connectivity map project produced by \code{\link{geneName2PerturbAnno}}}

\item{CMap_files}{a list of directories and urls produced by \code{\link{loadCMap}}}

\item{landmark_only}{read only landmark genes (rows, measured). Details: https://docs.google.com/document/d/1q2gciWRhVCAAnlvF2iRLuJ7whrGP6QjpsCMq1yWz7dU/edit}
}
\value{
object of class 'GCT' [package "cmapR"] with 7 slots containing z-score matrix, perturbation and feature details of the Connectivity map project
}
\description{
\code{readCMAP} reads Connectivity Map data given perturbation annotations for a set of HUGO gene names (or Entrez gene ID) including additional filtering by cell line, perturbation type (compound, shRNA, overexpression, e.g.) and time. Details \code{\link{geneName2PerturbAnno}}
}
\examples{
# read P53 knockdown perturbations
CMap_files = loadCMap(directory = "../regulatory_networks_by_cmap/data/cmap/")
p53IDs = geneName2PerturbAnno(CMap_files = CMap_files, is_touchstone = T, pert_types = "trt_sh.cgs")
CMAP = readCMAP(PerturbAnno = p53IDs, CMap_files = CMap_files)
# read all knockdown perturbations (3.4GB)
allIDs = geneName2PerturbAnno(gene_names = "all", CMap_files = CMap_files, is_touchstone = T, pert_types = "trt_sh.cgs")
CMAP = readCMAP(PerturbAnno = allIDs, CMap_files = CMap_files)
}
\seealso{
\code{\link{openCellInfo}}, \code{\link{loadCMap}}, \code{\link{perturbTable}}
}
\author{
Vitalii Kleshchevnikov
}
