% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removePCfromCMAP.R
\name{removePCfromCMAP}
\alias{removePCfromCMAP}
\title{Calculate and remove principal components from Connectivity map data}
\usage{
removePCfromCMAP(CMap_files, keep_PCs = "all", return_PCs = T,
  return_CMAP = T, save_CMAP = F, pert_types = c("trt_sh.cgs", "trt_oe",
  "trt_cp", "trt_lig", "trt_oe.mut", "trt_xpr", "trt_sh.css", "ctl_vehicle.cns",
  "ctl_vector.cns", "ctl_untrt.cns"))
}
\arguments{
\item{CMap_files}{a list of directories and urls produced by \code{\link{loadCMap}}}

\item{keep_PCs}{integer, indexes - which principal components to keep}

\item{return_PCs}{logical, return the output of \link[base]{svd} decomposition}

\item{return_CMAP}{logical, return Connectivity map data reconstructed using only specified PCs}

\item{save_CMAP}{logical, save Connectivity map data reconstructed using only specified PCs. Returned \code{CMap_files} contain the path to reconstructed Connectivity map data matrix and corresponding supplementary files.}

\item{pert_types}{a character vector of perturbation types, query for available perturbation types using \code{perturbTable(CMap_files, ~ pert_type)}. Details: https://docs.google.com/document/d/1q2gciWRhVCAAnlvF2iRLuJ7whrGP6QjpsCMq1yWz7dU/edit#}
}
\value{
\code{removePCfromCMAP}: data.table containing gene set id (single), which genes may regulate this set, test statistic and difference in medians between each gene set and all other genes
}
\description{
\code{removePCfromCMAP}: calculate and remove principal components from Connectivity map data. If CMAP data matrix have already been reconstructed using the same PCs and pert_types this function will read previously saved SVD output and reconstructed CMAP data matrix
}
