% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFtargetsINcmap.R
\name{showLargestRegulons}
\alias{showLargestRegulons}
\title{Show largest TF regulons}
\usage{
showLargestRegulons(regulons, N = 40)
}
\arguments{
\item{regulons}{data.table containing TF regulons (TF and target columns)}

\item{N}{number of largest regulons to display}
}
\description{
Show largest TF regulons
}
