% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdCMAP.R
\name{reconstructCMAP}
\alias{reconstructCMAP}
\alias{svdCMAP}
\title{Calculate and remove principal components from Connectivity map data}
\usage{
reconstructCMAP(CMap_files, remove_PCs = "none", return_PCs = T,
  return_CMAP = T, save_CMAP = F, pert_types = c("trt_sh.cgs", "trt_oe",
  "trt_cp", "trt_lig", "trt_oe.mut", "trt_xpr", "trt_sh.css", "ctl_vehicle.cns",
  "ctl_vector.cns", "ctl_untrt.cns"))

svdCMAP(CMap_files, return_PCs = T, is_touchstone = "all",
  pert_types = c("trt_sh.cgs", "trt_oe", "trt_cp", "trt_lig", "trt_oe.mut",
  "trt_xpr", "trt_sh.css", "ctl_vehicle.cns", "ctl_vector.cns",
  "ctl_untrt.cns"), pert_itimes = c("all"), cell_ids = c("all"),
  gene_names = c("all"), keep_one_oe = "all", landmark_only = F)
}
\arguments{
\item{CMap_files}{a list of directories and urls produced by \code{\link{loadCMap}}}

\item{remove_PCs}{integer or character, indexes - which principal components to remove. "none" to keep all PCs}

\item{return_PCs}{logical, return the output of \link[base]{svd} decomposition}

\item{return_CMAP}{logical, return Connectivity map data reconstructed using only specified PCs}

\item{save_CMAP}{logical, save Connectivity map data reconstructed using only specified PCs. Returned \code{CMap_files} contain the path to reconstructed Connectivity map data matrix and corresponding supplementary files.}

\item{pert_types}{a character vector of perturbation types, query for available perturbation types using \code{perturbTable(CMap_files, ~ pert_type)}. Details: https://docs.google.com/document/d/1q2gciWRhVCAAnlvF2iRLuJ7whrGP6QjpsCMq1yWz7dU/edit#}

\item{is_touchstone}{logical, select only the perturbation data that is a part of the touchstone dataset (genes profiled across all 9 core cell lines)? Details: https://docs.google.com/document/d/1q2gciWRhVCAAnlvF2iRLuJ7whrGP6QjpsCMq1yWz7dU/}

\item{pert_itimes}{a character vector of perturbation times, query for available perturbation times using \code{perturbTable(CMap_files, ~ pert_itime)}}

\item{cell_ids}{a character vector of cell line names, query for available cell line names using \code{perturbTable(CMap_files, ~ cell_id)}}

\item{gene_names}{a character vector of HUGO gene names. Use \code{"all"} to select all perturbations.}

\item{keep_one_oe}{keep only one overexpression experiment per gene and condition. Applicable only when \code{pert_types = "trt_oe"}. Perturbations where sig_id matches pert_id are retained ("one"). To invert the selection use "other". To select all use "all"}

\item{landmark_only}{read only landmark genes (rows, measured). Details: https://docs.google.com/document/d/1q2gciWRhVCAAnlvF2iRLuJ7whrGP6QjpsCMq1yWz7dU/edit}

\item{CMap_files}{a list of directories and urls produced by \code{\link{loadCMap}}}

\item{return_PCs}{logical, return the output of \link[base]{svd} decomposition. Alternatively just calculate and save.}

\item{pert_types}{a character vector of perturbation types, query for available perturbation types using \code{perturbTable(CMap_files, ~ pert_type)}. Details: https://docs.google.com/document/d/1q2gciWRhVCAAnlvF2iRLuJ7whrGP6QjpsCMq1yWz7dU/edit#}
}
\value{
\code{reconstructCMAP}:

\code{svdCMAP}:
}
\description{
\code{reconstructCMAP}: calculate and remove principal components from Connectivity map data. If CMAP data matrix have already been reconstructed using the same PCs and pert_types this function will read previously saved SVD output and reconstructed CMAP data matrix

\code{svdCMAP}: calculate and remove principal components from Connectivity map data. If CMAP data matrix have already been reconstructed using the same PCs and pert_types this function will read previously saved SVD output and reconstructed CMAP data matrix
}
\examples{
library(regNETcmap)
CMap_files = loadCMap(directory = "../regulatory_networks_by_cmap/data/cmap/")
library(regNETcmap)
CMap_files = loadCMap(directory = "../regulatory_networks_by_cmap/data/cmap/")
}
