% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFtargetsINcmap.R
\name{TFtargetsINcmap}
\alias{TFtargetsINcmap}
\title{Summarise TF targets enrichment when TF perturbed in Connectivity Map}
\usage{
TFtargetsINcmap(regulons, alternative = "less", pert_types = c("trt_oe"),
  cell_ids = c("A375", "A549", "HA1E", "HCC515", "HEPG2", "HT29", "MCF7",
  "PC3"), gene_names = c("NFKB1", "E2F1", "MYC", "TP53", "YY1", "JUN",
  "STAT1", "STAT3"), pert_itimes = c("96 h"), CMap_files,
  keep_one_oe = c("one", "other", "all")[1], landmark_only = F,
  clustermq_memory = 2000, clustermq_job_size = 1,
  min_targets_in_cmap = 3, force_gene_names = F)
}
\arguments{
\item{regulons}{data.table containing TF regulons (TF and target columns)}

\item{alternative}{indicates alternative hypothesis for \link[stats]{ks.test}. "trt_oe": negative_regulator = "less", positive_regulator = "greater"; "trt_sh.cgs" or anything else: negative_regulator = "greater"; positive_regulator = "less"}

\item{pert_types}{a character vector of perturbation types, query for available perturbation types using \code{perturbTable(CMap_files, ~ pert_type)}}

\item{cell_ids}{a character vector of cell line names, query for available cell line names using \code{perturbTable(CMap_files, ~ cell_id)}}

\item{gene_names}{a character vector of HUGO gene names. Use \code{"all"} to select all perturbations.}

\item{pert_itimes}{a character vector of perturbation times, query for available perturbation times using \code{perturbTable(CMap_files, ~ pert_itime)}}

\item{CMap_files}{a list of directories and urls produced by \code{\link{loadCMap}}}

\item{keep_one_oe}{keep only one overexpression experiment per gene and condition. Applicable only when \code{pert_types = "trt_oe"}. Perturbations where sig_id matches pert_id are retained ("one"). To invert the selection use "other". To select all use "all"}

\item{landmark_only}{look only at landmark genes. Details: https://docs.google.com/document/d/1q2gciWRhVCAAnlvF2iRLuJ7whrGP6QjpsCMq1yWz7dU/edit}

\item{clustermq_memory}{When using clustermq: memory requested for each job}

\item{clustermq_job_size}{When using clustermq: The number of function calls per job}

\item{min_targets_in_cmap}{minimal number of TF targets in cmap}

\item{force_gene_names}{look at all \code{gene_names} but less cell lines}
}
\value{
list containing data.table summary of TF perturbation effects on TF targets, object of class 'GCT' (CMap data), summary of genes vs cell lines available in CMap and returned (gene_cell_counts)
}
\description{
\code{TFtargetsINcmap}
}
\seealso{
\code{\link{regFromCMAP}}, \code{\link{completeCellGeneCombs}}, \code{\link{showLargestRegulons}}
}
\author{
Vitalii Kleshchevnikov
}
