% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneName2PerturbAnno.R, R/keep1OE.R
\name{geneName2PerturbAnno}
\alias{geneName2PerturbAnno}
\alias{anyGeneID2PerturbAnno}
\alias{keep1OE}
\title{Get Connectivity Map perturbation annotations for genes of interest, cell line and perturbation types and times}
\usage{
geneName2PerturbAnno(gene_names = "TP53", CMap_files,
  is_touchstone = c("all", T, F), pert_types = c("trt_sh.cgs", "trt_oe"),
  pert_itimes = c("all"), cell_ids = c("all"))

anyGeneID2PerturbAnno(keys = "Q9NZL3", CMap_files, org.db = c("UniProt.ws",
  "Homo.sapiens")[1], keytype = "UNIPROTKB", is_touchstone = c("all", T, F),
  pert_types = c("trt_sh.cgs", "trt_sh", "trt_sh.css"),
  pert_itimes = c("all"), cell_ids = c("all"))

keep1OE(PerturbAnno, keep_one_oe = c("one", "other", "all")[1],
  pert_types = "trt_oe", CMap_files)
}
\arguments{
\item{gene_names}{a character vector of HUGO gene names or other valid pert_iname. Use \code{"all"} to select all perturbations.}

\item{CMap_files}{a list of directories and urls produced by \code{\link{loadCMap}}}

\item{is_touchstone}{logical, select only the perturbation data that is a part of the touchstone dataset (genes profiled across all 9 core cell lines)? Details: https://docs.google.com/document/d/1q2gciWRhVCAAnlvF2iRLuJ7whrGP6QjpsCMq1yWz7dU/}

\item{pert_types}{a character vector of perturbation types, query for available perturbation types using \code{perturbTable(CMap_files, ~ pert_type)}}

\item{pert_itimes}{a character vector of perturbation times, query for available perturbation times using \code{perturbTable(CMap_files, ~ pert_itime)}}

\item{cell_ids}{a character vector of cell line names, query for available cell line names using \code{perturbTable(CMap_files, ~ cell_id)}}

\item{keys}{a character vector of identifiers, check possible options for type of identifier by calling keytypes(org.db)}

\item{org.db}{org.db of choice: Homo.sapiens or UniProt.ws::UniProt.ws(9606)}

\item{keytype}{type of identifier used in \code{keys}}

\item{PerturbAnno}{data.table containing the perturbation details from the Connectivity map project produced by \code{\link{geneName2PerturbAnno}}}

\item{keep_one_oe}{keep only one overexpression experiment per gene and condition. Applicable only when \code{pert_types = "trt_oe"}. Perturbations where sig_id matches pert_id are retained ("one"). To invert the selection use "other". To select all use "all"}

\item{pert_types}{find duplicates in this perturbation type}

\item{CMap_files}{a list of directories and urls produced by \code{\link{loadCMap}}}
}
\value{
data.table containing the perturbation details from the Connectivity map project

data.table containing the perturbation details from the Connectivity map project

data.table containing the perturbation details from the Connectivity map project. Only one overexpression experiment per gene and condition is retained.
}
\description{
\code{geneName2PerturbAnno} retrieves perturbation annotations for a set of HUGO gene names (or Entrez gene ID) including additional filtering by cell line, perturbation type (compound, shRNA, overexpression, e.g.) and time

\code{anyGeneID2PerturbAnno} retrieves perturbation annotations for a set of gene idenfified using any type of ID in Homo.sapiens or UniProt.ws, including additional filtering by cell line, perturbation type (compound, shRNA, overexpression, e.g.) and time

\code{keep1OE} Keep only 1 overexpression experiment per gene and condition. Designed to be applicable to \code{pert_types = "trt_oe"}.
}
\details{
is_touchstone: A boolean indicating whether the corresponding signature or perturbagen is a member of the Touchstone dataset. Touchstone is a term applied to the subset of CMap perturbagens that are well-annotated and that were systematically profiled across the majority of the core set of 9 cell lines at standardized conditions. Because of these properties Touchstone dataset well-suited as a reference compendium against with to compare external queries.
}
\seealso{
\code{\link{openCellInfo}}, \code{\link{loadCMap}}, \code{\link{perturbTable}}

\code{\link{openCellInfo}}, \code{\link{loadCMap}}, \code{\link{perturbTable}}
}
\author{
Vitalii Kleshchevnikov

Vitalii Kleshchevnikov

Vitalii Kleshchevnikov
}
