% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCMap.R
\name{openCellInfo}
\alias{openCellInfo}
\alias{openFeatureData}
\alias{openpData}
\alias{openPerturbDetails}
\title{Read connectivity map annotations}
\usage{
openCellInfo(CMap_files)

openFeatureData(CMap_files)

openpData(CMap_files)

openPerturbDetails(CMap_files)
}
\arguments{
\item{CMap_files}{a list of directories and urls produced by \code{\link{loadCMap}}}

\item{CMap_files}{a list of directories and urls produced by \code{\link{loadCMap}}}

\item{CMap_files}{a list of directories and urls produced by \code{\link{loadCMap}}}

\item{CMap_files}{a list of directories and urls produced by \code{\link{loadCMap}}}
}
\value{
data.table containing Cell line Information from the Connectivity map project

data.table containing featureData (\code{\link[Biobase]{featureData}}) from the Connectivity map project

data.table containing the sample descriptions (pData) (\code{\link[Biobase]{pData}}) from the Connectivity map project

data.table containing the perturbation details from the Connectivity map project
}
\description{
\code{openCellInfo} reads Connectivity Map project cell line data

\code{openFeatureData} reads Connectivity Map project featureData

\code{openpData} read Connectivity Map project sample descriptions (pData)

\code{openPerturbDetails} read Connectivity Map project perturbation details
}
\author{
Vitalii Kleshchevnikov

Vitalii Kleshchevnikov

Vitalii Kleshchevnikov

Vitalii Kleshchevnikov
}
