% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFtargetsINcmap.R
\name{plotTFtargetsINcmap}
\alias{plotTFtargetsINcmap}
\title{Plot TF perturbation effects (CMap) on TF targets}
\usage{
plotTFtargetsINcmap(res, TFsels = NULL,
  title = "TF that positively regulate their targets \\n(shRNA induces shift to the left)")
}
\arguments{
\item{res}{data.table containing summary of TF perturbation effects on TF targets returned by \code{\link{TFtargetsINcmap}}}

\item{TFsels}{which TF to plot, defaults to all in \code{res}}
}
\description{
Plot TF perturbation effects (CMap) on TF targets
}
