% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFtargetsINcmap.R
\name{plotTFtargetsINcmap}
\alias{plotTFtargetsINcmap}
\title{Plot TF perturbation effects (CMap) on TF targets}
\usage{
plotTFtargetsINcmap(res, TFsels = NULL,
  title = "TF that positively regulate their targets \\n(shRNA induces shift to the left)",
  lab_text_size = 2, lab_text_y = 1.2, lab_text_x = -2.2,
  strip.text_size = 8, vline_size = 1, xlim = c(-3.3, 3.3),
  ylim = c(0, 1.7))
}
\arguments{
\item{res}{data.table containing summary of TF perturbation effects on TF targets returned by \code{\link{TFtargetsINcmap}}}

\item{TFsels}{which TF to plot, defaults to all in \code{res}}

\item{xlim}{passed to \link[ggplot2]{xlim}}
}
\description{
Plot TF perturbation effects (CMap) on TF targets
}
