% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCMAP.R
\name{readCMAP}
\alias{readCMAP}
\title{Read Connectivity Map data given perturbation annotations}
\usage{
readCMAP(PerturbAnno, CMap_files)
}
\arguments{
\item{PerturbAnno}{a character vector of HUGO gene names}

\item{CMap_files}{a list of directories and urls produced by \code{\link{loadCMap}}}
}
\value{
data.table containing the perturbation details from the Connectivity map project
}
\description{
\code{readCMAP} reads Connectivity Map data given perturbation annotations for a set of HUGO gene names (or Entrez gene ID) including additional filtering by cell line, perturbation type (compound, shRNA, overexpression, e.g.) and time. Details \code{\link{geneName2PerturbAnno}}
}
\examples{
# read P53 knockdown perturbations
CMap_files = loadCMap(directory = "../regulatory_networks_by_cmap/data/cmap/")
p53IDs = geneName2PerturbAnno(CMap_files = CMap_files, is_touchstone = T, pert_types = "trt_sh.cgs")
CMAP = readCMAP(PerturbAnno = p53IDs, CMap_files = CMap_files)
# read all knockdown perturbations (3GB)
fdata = openFeatureData(CMap_files)
allIDs = geneName2PerturbAnno(gene_names = unique(fdata$pr_gene_symbol), CMap_files = CMap_files, is_touchstone = T, pert_types = "trt_sh.cgs")
CMAP = readCMAP(PerturbAnno = allIDs, CMap_files = CMap_files)
}
\seealso{
\code{\link{openCellInfo}}, \code{\link{loadCMap}}, \code{\link{perturbTable}}
}
\author{
Vitalii Kleshchevnikov
}
